support decimal input

+++ components/SetLimitModal.js
@@ -28,7 +28,7 @@ export default function SetLimitModal({
                         <TextInput
                             style={styles.limitInput}
                             defaultValue={limit.toString()}
-                            onChangeText={(text) => text ? setValue(parseInt(text)) : setValue(0)}
+                            onChangeText={(text) => text ? setValue(parseFloat(text)) : setValue(0)}
                             keyboardType='number-pad'
                             returnKeyType='done'
                         />
